requirements.txt
googlesearch-python
requests
colorama
